[manager]feature: refactor DispatchAlarm
