+++ manager/src/main/java/com/usthe/manager/component/alerter/impl/WebHookAlertNotifyHandlerImpl.java
@@ -1,0 +1,44 @@
+package com.usthe.manager.component.alerter.impl;
+
+import com.usthe.common.entity.alerter.Alert;
+import com.usthe.common.entity.manager.NoticeReceiver;
+import com.usthe.manager.component.alerter.AlertNotifyHandler;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Component;
+import org.springframework.web.client.ResourceAccessException;
+import org.springframework.web.client.RestTemplate;
+
+/**
+ * @author <a href="mailto:Musk.Chen@fanruan.com">Musk.Chen</a>
+ * @since 2022/4/24
+ */
+@Component
+@RequiredArgsConstructor
+@Slf4j
+final class WebHookAlertNotifyHandlerImpl implements AlertNotifyHandler {
+    private final RestTemplate restTemplate;
+
+    @Override
+    public void send(NoticeReceiver receiver, Alert alert) {
+        try {
+            ResponseEntity<String> entity = restTemplate.postForEntity(receiver.getHookUrl(), alert, String.class);
+            if (entity.getStatusCode().value() < HttpStatus.BAD_REQUEST.value()) {
+                log.debug("Send WebHook: {} Success", receiver.getHookUrl());
+            } else {
+                log.warn("Send WebHook: {} Failed", receiver.getHookUrl());
+            }
+        } catch (ResourceAccessException e) {
+            log.warn("Send WebHook: {} Failed: {}.", receiver.getHookUrl(), e.getMessage());
+        } catch (Exception e) {
+            log.error(e.getMessage(), e);
+        }
+    }
+
+    @Override
+    public byte type() {
+        return 2;
+    }
+}
